+++ TimeTableManagementSystemNew/WorkingDaysandHours.cs
@@ -156,19 +156,22 @@ namespace TimeTableManagementSystemNew
         {
             if (timeslotID > 0)
             {
-                SqlCommand cmd = new SqlCommand("DELETE FROM No_of_Working_Days WHERE WorkID=@ID", con);
-                cmd.CommandType = CommandType.Text;
+                if (MessageBox.Show("Are you sure to delete?", "Delete Record", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                {
+                    SqlCommand cmd = new SqlCommand("DELETE FROM No_of_Working_Days WHERE WorkID=@ID", con);
+                    cmd.CommandType = CommandType.Text;
 
-                cmd.Parameters.AddWithValue("@ID", this.timeslotID);
-                con.Open();
-                cmd.ExecuteNonQuery();
-                con.Close();
+                    cmd.Parameters.AddWithValue("@ID", this.timeslotID);
+                    con.Open();
+                    cmd.ExecuteNonQuery();
+                    con.Close();
 
-                MessageBox.Show("Successfully Deleted Working Days and Hours", "Saved", MessageBoxButtons.OK, MessageBoxIcon.Information);
-                GetTimeSlotRecord();
+                   
+                    GetTimeSlotRecord();
 
 
-                ResetValue();
+                    ResetValue();
+                }
             }
             else
             {
